Add files via upload
Completion changes:
Made the robot check the left top and right parts of the image.
Added the white or not white to an array that contains all the left, top, and right checked pixels and then check for grouped whiteness. Calculates the average position of the line. If it differs from the center of the array (Center is going straight) then it turns the corresponding motor accordingly does not fully work though needs imporvement